Making OpenCV and Port Audio optional

+++ interfaceMicrophone/include/interfaceMicrophone.h
@@ -6,27 +6,27 @@
  *
  *  @author Aashish Sheshadri
  *  @author Rohit Harchandani
- *  
+ *
  *  The MIT License (MIT)
- *  
+ *
  *  Copyright (c) 2015 PayPal
- *  
+ *
  *  Permission is hereby granted, free of charge, to any person obtaining a copy
- *  of this software and associated documentation files (the "Software"), to 
- *  deal in the Software without restriction, including without limitation the 
- *  rights to use, copy, modify, merge, publish, distribute, sublicense, and/or 
+ *  of this software and associated documentation files (the "Software"), to
+ *  deal in the Software without restriction, including without limitation the
+ *  rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
  *  sell copies of the Software, and to permit persons to whom the Software is
  *  furnished to do so, subject to the following conditions:
- *  
+ *
  *  The above copyright notice and this permission notice shall be included in
  *  all copies or substantial portions of the Software.
- *  
+ *
  *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
  *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING 
- *  FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER  
+ *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+ *  FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  *  DEALINGS IN THE SOFTWARE.
  */
 
@@ -299,9 +299,9 @@ void InterfaceMicrophone::int16toBytes(II begin, II end, OI out) {
 	// Loop through int16 stream.
 	while (begin != end) {
 		// Convert int16 to bytes and load them into out.
-		*out = static_cast<uint8_t>(*begin & 0x00FF);
+		*out = static_cast<uint8_t>(*begin & uint16_t(0x00FF));
 		++out;
-		*out = static_cast<uint8_t>((*begin & 0xFF00) >> 8);
+		*out = static_cast<uint8_t>((*begin & uint16_t(0xFF00)) >> 8);
 		++begin;
 	}
 }
